home work 5 double fix

+++ src/com/company/GameResult.java
@@ -3,5 +3,5 @@ package com.company;
 public class GameResult {
     String userName;
     int attempts;
-    long time;
+    double time;
 }
